Modais prontas com js e correções nas inserções de senhas

+++ set_new_password_page.php
@@ -13,7 +13,7 @@
     $email= $_GET['email'];
     $key = $_GET['confirmation'];
 
-    $_SESSION["msg"] = "";
+    $_SESSION["msg_error"] = "";
 
     $check_key = checkKey($email, $key, $pdo);
 
@@ -23,21 +23,27 @@
     }
 
     if(isset($_POST['email'])){
-        if($check_key){   
-            $password=password_hash($_POST['newPassword'], PASSWORD_DEFAULT);
-            $array = [$password, $email];
-            $update_password = updatePassword($array, $pdo);
+        if($check_key){
+            if($_POST['newPassword'] == $_POST['newPasswordConfirmation']){
+                $password=password_hash($_POST['newPassword'], PASSWORD_DEFAULT);
+                $array = [$password, $email];
+                $update_password = updatePassword($array, $pdo);
+        
+                if($update_password){
+                    deleteRequest($email, $key, $pdo);
     
-            if($update_password){
-                deleteRequest($email, $key, $pdo);
-
-                $_SESSION["senha_atualizada"] = "Senha atualizada com sucesso!";
-
-                header("Location:login.php");
+                    $_SESSION["senha_atualizada"] = "Senha atualizada com sucesso!";
+    
+                    header("Location:login.php");
+                    exit;
+                }else{
+                    $_SESSION["senha_atualizada_error"] = "Erro ao atualizar a senha.";
+    
+                    header("Location:login.php");
+                    exit;
+                }
             }else{
-                $_SESSION["senha_atualizada_error"] = "Erro ao atualizar a senha";
-
-                header("Location:login.php");
+                $_SESSION["msg_error"] = "Senhas não conferem.";
             }
             
         }
@@ -54,26 +60,32 @@
     <main>
         <section class="position-forms">
             <div class="content-box" >
-                <form action="set_new_password_page.php?email=<?php echo($email) ?>&confirmation=<?php echo($key) ?> " method="POST">
+                <form action="set_new_password_page.php?email=<?php echo($email) ?>&confirmation=<?php echo($key) ?> " onsubmit="return validaCreateAccount()" method="POST">
                     <h2>Nova senha:</h2>
                     <div class="form-container">
                         <div class="input-container">
                             <label for="newPassword">Nova senha:</label>
-                            <input type="password" id="newPassword" name="newPassword" placeholder="********" autocomplete="off">
+                            <input type="password" id="password" name="newPassword" placeholder="********" autocomplete="off">
                             <input type="hidden" id="email" name="email" value='<?php echo $email ?>'>
                             <input type="hidden" id="confirmation" name="confirmation" value='<?php echo $key ?>'>
                         </div>
                         <div class="input-container">
                             <label for="PasswordConfirmation">Confirme sua senha</label>
-                            <input type="password" id="PasswordConfirmation" name="PasswordConfirmation" placeholder="********" autocomplete="off" required>
+                            <input type="password" id="PasswordConfirmation" name="newPasswordConfirmation" placeholder="********" autocomplete="off" required>
                         </div>
                     </div>
                     
                     <button type="submit" class='btn-submit'>
                         Criar conta!
                     </button>
+                    <div class='msg'>
+                        <span class="msg-error">
+                            <?php echo ($_SESSION["msg_error"]); ?>
+                        </span>
+                    </div>
                     
                 </form>
+
             </div>
         </section>
     </main>
